requirements.txt
Pillow==8.2.0
qrcode==6.1
python-barcode==0.13.1
certifi==2021.5.30
chardet==4.0.0
idna==2.10
requests==2.25.1
urllib3==1.26.5
